[CLIENT] Restructure & Extract Components
 * Extract tables from pages to their separate components.
 * Create 3 components modules to hold related components for pages.
 * Refactor some types/code.
