+++ client/pages/certification/edit/[id].tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useRouter } from "next/dist/client/router";
 
-import { CertificationForm } from "../../../components/forms/CertificationForm";
+import { CertificationForm } from "../../../components/certifications/CertificationForm";
 import {
   useCertificationQuery,
   UserRole,
