+++ client/pages/employee-certification/edit/[id].tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useRouter } from "next/dist/client/router";
 
 import { Loading } from "../../../components/common/Loading";
-import { EmployeeCertificationForm } from "../../../components/forms/EmployeeCertificationForm";
+import { EmployeeCertificationForm } from "../../../components/certifications/EmployeeCertificationForm";
 import {
   EmployeeCertificationInput,
   useEmployeeCertificationQuery,
@@ -45,7 +45,7 @@ const EditEmployeeCertification: React.FC<{}> = () => {
 
   return (
     <EmployeeCertificationForm
-      intitalCertitficationProviderId={
+      initialCertificationProviderId={
         employeeCertification.certification.certificationProvider
           .certificationProviderId
       }
