+++ client/pages/release-request/edit/[id].tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useRouter } from "next/dist/client/router";
 
 import { Loading } from "../../../components/common/Loading";
-import { ReleaseRequestForm } from "../../../components/forms/ReleaseRequestForm";
+import { ReleaseRequestForm } from "../../../components/requests/ReleaseRequestForm";
 import {
   useReleaseRequestQuery,
   UserRole,
