+++ client/pages/resource-request/create.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useRouter } from "next/dist/client/router";
 
-import { ResourceRequestForm } from "../../components/forms/ResourceRequestForm";
+import { ResourceRequestForm } from "../../components/requests/ResourceRequestForm";
 import {
   ResourceRequestInput,
   useCreateResourceRequestMutation,
